clustering_client
